add feature constructor in starlark
