+++ pkg/star/compile.go
@@ -15,7 +15,6 @@
 package star
 
 import (
-	"fmt"
 	"io/ioutil"
 	"os"
 
@@ -24,12 +23,6 @@ import (
 	"github.com/stripe/skycfg/go/protomodule"
 	"go.starlark.net/starlark"
 	"google.golang.org/protobuf/reflect/protoregistry"
-)
-
-const (
-	starVariableDefault     = "default"
-	starVariableDescription = "description"
-	starVariableRules       = "rules"
 )
 
 type Compiler interface {
@@ -54,37 +47,12 @@ func NewCompiler(registry *protoregistry.Types, protoDir, starfilePath, featureN
 	}
 }
 
-type variables struct {
-	defaultVal, description, rules starlark.Value
-}
-
 func (c *compiler) Compile() (*feature.Feature, error) {
 	// Execute the starlark file to retrieve its contents (globals)
-	vars, err := loadVariables(c.registry, c.starfilePath)
-	if err != nil {
-		return nil, errors.Wrap(err, "load globals")
-	}
-	// Now, get the config values and turn them into our internal representation (feature.Feature).
-	f, err := getFeature(vars.defaultVal)
-	if err != nil {
-		return nil, errors.Wrap(err, "get feature")
-	}
-	f.Key = c.featureName
-	f.Description, err = getDescription(vars.description)
-	if err != nil {
-		return nil, errors.Wrap(err, "get description")
-	}
-	if err = addRules(f, vars.rules); err != nil {
-		return nil, errors.Wrap(err, "add rules")
-	}
-	return f, nil
-}
-
-func loadVariables(registry *protoregistry.Types, starfilePath string) (*variables, error) {
 	thread := &starlark.Thread{
 		Name: "load",
 	}
-	reader, err := os.Open(starfilePath)
+	reader, err := os.Open(c.starfilePath)
 	if err != nil {
 		return nil, errors.Wrap(err, "open starfile")
 	}
@@ -93,125 +61,18 @@ func loadVariables(registry *protoregistry.Types, starfilePath string) (*variabl
 	if err != nil {
 		return nil, errors.Wrap(err, "read starfile")
 	}
-	protoModule := protomodule.NewModule(registry)
-	globals, err := starlark.ExecFile(thread, starfilePath, moduleSource, starlark.StringDict{
-		"proto": protoModule,
+	protoModule := protomodule.NewModule(c.registry)
+	globals, err := starlark.ExecFile(thread, c.starfilePath, moduleSource, starlark.StringDict{
+		"proto":   protoModule,
+		"feature": starlark.NewBuiltin("feature", makeFeature),
 	})
 	if err != nil {
 		return nil, errors.Wrap(err, "starlark execfile")
 	}
-	vars := &variables{}
-	var ok bool
-	vars.defaultVal, ok = globals[starVariableDefault]
-	if !ok {
-		return nil, fmt.Errorf("required variable %s not found", starVariableDefault)
+	f, err := newFeatureBuilder(globals).build()
+	if err != nil {
+		return nil, errors.Wrap(err, "build")
 	}
-	vars.description, ok = globals[starVariableDescription]
-	if !ok {
-		return nil, fmt.Errorf("required variable %s not found", starVariableDescription)
-	}
-	vars.rules, ok = globals[starVariableRules]
-	if !ok {
-		vars.rules = starlark.None
-	}
-	return vars, nil
-}
-
-func getFeature(defaultVal starlark.Value) (*feature.Feature, error) {
-	// check if this is a complex type
-	message, ok := protomodule.AsProtoMessage(defaultVal)
-	if ok {
-		return feature.NewComplexFeature(message), nil
-	}
-	// check if this is a supported primitive type
-	switch typedVal := defaultVal.(type) {
-	case starlark.Bool:
-		return feature.NewBoolFeature(bool(typedVal)), nil
-	default:
-		return nil, fmt.Errorf("received default value with unsupported type %T", typedVal)
-	}
-}
-
-func getDescription(descriptionVal starlark.Value) (string, error) {
-	dsc, ok := descriptionVal.(starlark.String)
-	if !ok {
-		return "", fmt.Errorf("description must be a string (got a %s)", descriptionVal.Type())
-	}
-	return dsc.GoString(), nil
-}
-
-func addRules(f *feature.Feature, rulesVal starlark.Value) error {
-	if _, ok := rulesVal.(starlark.NoneType); ok {
-		// no rules provided, continue
-		return nil
-	}
-	seq, ok := rulesVal.(starlark.Sequence)
-	if !ok {
-		return fmt.Errorf("rules: did not get back a starlark sequence: %v", rulesVal)
-	}
-	it := seq.Iterate()
-	defer it.Done()
-	var val starlark.Value
-	var i int
-	for it.Next(&val) {
-		if val == starlark.None {
-			return fmt.Errorf("type error: [%v] %v", val.Type(), val.String())
-		}
-		tuple, ok := val.(starlark.Tuple)
-		if !ok {
-			return fmt.Errorf("type error: expecting tuple, got %v", val.Type())
-		}
-		if tuple.Len() != 2 {
-			return fmt.Errorf("expecting tuple of length 2, got length %d: %v", tuple.Len(), tuple)
-		}
-		conditionStr, ok := tuple.Index(0).(starlark.String)
-		if !ok {
-			return fmt.Errorf("type error: expecting string, got %v: %v", tuple.Index(0).Type(), tuple.Index(0))
-		}
-		// TODO: parse into ruleslang
-		if conditionStr.GoString() == "" {
-			return fmt.Errorf("expecting valid ruleslang, got %s", conditionStr.GoString())
-		}
-		ruleVal := tuple.Index(1)
-		switch f.FeatureType {
-		case feature.FeatureTypeComplex:
-			message, ok := protomodule.AsProtoMessage(ruleVal)
-			if !ok {
-				return typeError(f.FeatureType, i, ruleVal)
-			}
-			f.Rules = append(f.Rules, &feature.Rule{
-				Condition: conditionStr.GoString(),
-				Value:     message,
-			})
-		case feature.FeatureTypeBool:
-			typedRuleVal, ok := ruleVal.(starlark.Bool)
-			if !ok {
-				return typeError(f.FeatureType, i, ruleVal)
-			}
-			f.Rules = append(f.Rules, &feature.Rule{
-				Condition: conditionStr.GoString(),
-				Value:     bool(typedRuleVal),
-			})
-		default:
-			return fmt.Errorf("unsupported type %s for rule #%d", f.FeatureType, i)
-		}
-		i++
-	}
-
-	return nil
-}
-
-func typeError(expectedType feature.FeatureType, ruleIdx int, value starlark.Value) error {
-	return fmt.Errorf("expecting %s for rule idx #%d, instead got %T", starType(expectedType), ruleIdx, value)
-}
-
-func starType(ft feature.FeatureType) string {
-	switch ft {
-	case feature.FeatureTypeComplex:
-		return "protoMessage"
-	case feature.FeatureTypeBool:
-		return fmt.Sprintf("%T", starlark.False)
-	default:
-		return "unknown"
-	}
+	f.Key = c.featureName
+	return f, nil
 }
